requirements.txt
# ntoebook (.ipybn)
wcwidth==0.2.13
ipykernel==6.29.5
ipywidgets==8.1.4
jupyterlab-widgets==3.0.12
widgetsnbextension==4.0.12

# general use components
pdoc==12.*
flake8==6.0.0
pytest==7.1.*
python-dotenv==1.1.1

# general & notebook (.ipynb)_
tqdm

# AI modules
torch
torchaudio
torchcodec
openai-whisper
ollama
transformers 

# Components puramente python
pandas
scipy
pydub
ffmpeg-python

# langchain==0.3.25
# langchain-core==0.3.62
# langchain-ollama==0.3.3
# langchain-text-splitters==0.3.8
# chatterbox-tts==0.1.1
# diffusers==0.29.0
# huggingface-hub==0.32.2